+++ CHANGELOG.md
@@ -12,7 +12,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 * Deferred Waveform render_to_image operations to synchronize.  Fixes #239.
 * Added app "opengl" setting and use "desktop" by default.
-
+* Fixed js220 device settings log warnings. 
 
 ## 1.0.42